requirements.txt
typer[all]

# dev
black
flake8
mkdocs-material
mkdocstrings-python
mypy
pytest-cov
snakeviz
